move(encoding): variables from unstable_base64.ts to _common64.ts
